Move engine property from Target to Drawable
