+++ src/scratch/drawable_p.h
@@ -5,9 +5,12 @@
 namespace libscratchcpp
 {
 
+class IEngine;
+
 struct DrawablePrivate
 {
         int layerOrder = 0;
+        IEngine *engine = nullptr;
 };
 
 } // namespace libscratchcpp
