+++ test/scratch_classes/CMakeLists.txt
@@ -35,7 +35,9 @@ add_executable(
 target_link_libraries(
   drawable_test
   GTest::gtest_main
+  GTest::gmock_main
   scratchcpp
+  scratchcpp_mocks
 )
 
 gtest_discover_tests(drawable_test)
