+++ test/scratch_classes/target_test.cpp
@@ -797,16 +797,6 @@ TEST(SpriteTest, BubbleTextRedraw)
     target.bubble()->setText("");
 }
 
-TEST(TargetTest, Engine)
-{
-    Target target;
-    ASSERT_EQ(target.engine(), nullptr);
-
-    EngineMock engine;
-    target.setEngine(&engine);
-    ASSERT_EQ(target.engine(), &engine);
-}
-
 TEST(TargetTest, DataSource)
 {
     TargetMock target;
